+++ Drawingtoon/Drawingtoon/Shared/ParameterEncoder.swift
@@ -1,0 +1,12 @@
+//
+//  ParameterEncoder.swift
+//  Drawingtoon
+//
+//  Created by Heejung Yang on 9/23/25.
+//
+
+import Foundation
+
+protocol ParameterEncoder {
+    func encode<Parameters: Encodable>(_ parameters: Parameters, into request: URLRequest) throws -> URLRequest
+}
